added Podcast episodes, articles, 2 blockchain courses
added small desc to each of them, check if the language is okay and make changes if required. More podcast episodes and articles can be added if required. However i don't think the 2 edx courses that i added should go with the what the blockchain segment. They are the best resources imo and should be displayed separately but you take the call.